Cache the calls into factor_ecm.
Originally, I avoided doing this because technically two ECM runs are
not the same unless forcing the same sigma values, but the way ECM is
used in this program, it doesn't really matter. We never call factor_ecm
again with the same number to run additional curves at the same B1
value.

If by some chance, a given number ends up being factored twice, it
doesn't matter if a given ECM level could have found a factor if run
again, because it'll ultimately follow the same cached path as the
previous time that number was factored.

Signed-off-by: Jason Lynch <jason@aexoden.com>