Improve the choice between ECM and NFS early in data colleciton.
First, we refactor the ECM code to only determine if more ECM is needed
once per number. This way we don't end up with it changing on multiple
calls to the routine.

Second, instead of just using the calculated probability of finding a
factor, we start with an assumed probability (that may be way off), and
move toward the true value as the number of samples increases. This
avoids some undesired behavior (doing NFS way too early) when no factors
have yet ever been found on a given level.

Signed-off-by: Jason Lynch <jason@aexoden.com>