Fix some ECM corner cases.
In certain situations, the number of samples for each ECM level are not
monotonically decreasing. (Levels that do fewer curves might get data
for several different configured thread counts, and lower levels may be
skipped if dealing with a cofactor.)

First, we assume each level has a 10% chance of a factor, and then use
an exponentially decaying average to move to the true value as the
number of samples increases. This avoids pathological behavior if there
is a single sample at a level that found a factor (which breaks
calculating average times).

Second, instead of only checking the number of samples at the level one
beyond the minimum time, we check all levels from the minimum to one
beyond the minimum time.

Signed-off-by: Jason Lynch <jason@aexoden.com>